English translation of st_057110
As with st_057060, st_057115 and st_057117 seem to be updated versions of the text in this file, although they are only for partial sections. st_057115 covers lines 12 to 21, and st_057117 lines 22 to 43. Any changes to those lines should be made in their respective files, or they likely won't display properly.