English translation of st_061015.csv
Line 64, dunno what to call 星銃, which I'm assuming is Harriet's genesis name. Or part of it? Maybe some kind of new weapon type? I'm just leaving it as 'a weapon' for now.
Also, line 78 is Lisa and Harriet speaking together, but I'm not sure if there's a way to express not wanting to do something in third person in a way that Harriet also would, so... maybe turn the 'I' into 'we'?