Translated some more greetings
Line 28: "名残惜しいね" was quite awkward to translate. "名残惜しい" means "regret (at parting)" and/or "reluctance (to part)". The -ね is a particle used to ask for a listener’s confirmation, therefore it can be translated as "right?" or "isn’t it?". If it feels off to you, feel free to find a better alternative.